+++ src/cert/mod.rs
@@ -587,10 +587,25 @@ impl ResourceCert {
     pub fn repository_uri(&self) -> Option<uri::Rsync> {
         self.cert.extensions.repository_uri()
     }
+    
+    /// Returns the signed object rsync URI of this certificate if available.
+    pub fn signed_object_uri(&self) -> Option<uri::Rsync> {
+        self.cert.extensions.signed_object_uri()
+    }
+
+    /// Returns a reference to the validity.
+    pub fn validity(&self) -> &Validity {
+        &self.cert.validity
+    }
 
     /// Returns information about the TAL this certificate is based on.
     pub fn tal(&self) -> &Arc<TalInfo> {
         &self.tal
+    }
+
+    /// Converts the certificate into its TAL info.
+    pub fn into_tal(self) -> Arc<TalInfo> {
+        self.tal
     }
 }
 
@@ -638,6 +653,14 @@ impl Validity {
 
     pub fn from_secs(secs: i64) -> Self {
         Self::from_duration(Duration::seconds(secs))
+    }
+
+    pub fn not_before(&self) -> Time {
+        self.not_before
+    }
+
+    pub fn not_after(&self) -> Time {
+        self.not_after
     }
 
     pub fn take_from<S: decode::Source>(